requirements.txt
anyascii==0.3.1
beautifulsoup4==4.11.1
bpemb==0.3.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
conllu==4.5.2
contourpy==1.0.6
cssselect==1.2.0
cycler==0.11.0
Deprecated==1.2.13
Distance==0.1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
flair==0.11.3
fonttools==4.38.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
huggingface-hub==0.11.0
hyperopt==0.2.7
idna==3.4
importlib-metadata==3.10.1
Janome==0.4.2
jieba==0.42.1
jieba3k==0.35.1
joblib==1.2.0
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lmdb==1.3.0
lxml==4.9.1
marisa-trie==0.7.8
matplotlib==3.6.2
more-itertools==9.0.0
mpld3==0.3
mwparserfromhell==0.6.4
networkx==2.8.8
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.4
overrides==3.1.0
packaging==21.3
pandas==1.5.1
pause==0.3
Pillow==9.3.0
pptree==3.1
py4j==0.10.9.7
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywikibot==7.7.1
PyYAML==6.0
regex==2022.10.31
REL @ git+https://github.com/informagi/REL@c38b595498b509a227e62c6b253628dcb69fe5de
requests==2.28.1
requests-file==1.5.1
scikit-learn==1.1.3
scipy==1.9.3
segtok==1.5.11
sentencepiece==0.1.95
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.2.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
tabulate==0.9.0
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
torch==1.13.0
torchaudio==2.0.0.dev20230219
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
waybackpy==3.0.6
wcwidth==0.2.5
Wikipedia-API==0.5.4
wikipedia2vec==1.0.5
wrapt==1.14.1
zipp==3.10.0
